+++ NEWS.md
@@ -7,6 +7,8 @@ BUGFIXES :
 * `.importOutput()` Fixed area outputs import for Antares ≥ 9.3 when output columns differ between areas.
 * `.readDistrictsDefAPI()` treats case when `apply-filter` property is not written.
 * `.getStartDate()` is case insensitive to treat property `first-month-in-year` 
+* `.importOutputForAreas()`from Antares 9.3, for values files, if select is not null, get all the columns and select the desired ones at the end.
+
 
 # antaresRead 2.9.3